chore: refactor hardware wallet init screens into own file
